MATERIAL screen and storing material to FLASH

+++ src/t14-flash.c
@@ -7,7 +7,7 @@
 #include "t14-flash.h"
 #include "stm32l1xx_flash.h"
 
-FLASH_Status FLASH_Write_DataWord(uint32_t address, uint8_t data) {
+FLASH_Status FLASH_Write_DataWord(uint32_t address, uint32_t data) {
   FLASH_Status status = FLASH_COMPLETE;
   address = address + 0x08080000;
   FLASH_UnlockData();  //Unprotect the EEPROM to allow writing
@@ -16,8 +16,8 @@ FLASH_Status FLASH_Write_DataWord(uint32_t address, uint8_t data) {
   return status;
 }
 
-uint8_t FLASH_Read_DataWord(uint32_t address) {
-  uint8_t tmp = 0;
+uint32_t FLASH_Read_DataWord(uint32_t address) {
+  uint32_t tmp = 0;
   address = address + 0x08080000;
   tmp = *(__IO uint32_t*)address;
   return tmp;
